Refactor into modular files with literate comments and bug fixes
Split the monolithic weatherstar.c into focused modules:
- fb.h: framebuffer types, palette, drawing primitives
- font.h: 5x7 bitmap font data and text rendering
- icons.h: weather icon drawing (sun, cloud)
- display.h: WeatherStar 4000 screen layout composition
- screenshot.h: PNG output (libpng) and ANSI terminal output
- main.c: entry point and CLI parsing

Still compiles as a single translation unit (gcc main.c) — no
multi-object build complexity.

Remove redundant weatherstar.sh wrapper; Makefile is sufficient.

Bug fixes and improvements:
- Fix memory leak: row buffer leaked on libpng longjmp error path
  (allocate before setjmp so it can be freed in the error handler)
- Fix rounded-rect corner clearing: use integer distance² comparison
  instead of sqrtf — faster and avoids float-to-int edge cases
- Centralise background gradient into fb_bg_at() helper to eliminate
  duplicated interpolation code across 4 corner branches
- Buffer ANSI output per-line instead of per-pixel printf (10x fewer
  syscalls, ~3.6 MB output now written in 200 fwrite calls vs 128k)
- Add null checks for png_create_write_struct / png_create_info_struct
